Work Batch 1
Reworked the pack importer to support pack2 files
added handling for a namelist stage during the import
added a zlib decompression stage for compressed assets
added handling for incorrectly converted unsigned to signed values
modified mesh importer to support larger meshes. Still work in progress, need an interface layer to convert a ulong read command into a bunch of uint read commands

+++ PS2LS/ps2ls/Forms/MaterialBrowser.cs
@@ -79,7 +79,7 @@ namespace ps2ls.Forms
 
         private void update()
         {
-            glControl1.Camera.AspectRatio = (Single)glControl1.ClientSize.Width / (Single)glControl1.ClientSize.Height;
+            glControl1.Camera.AspectRatio = glControl1.ClientSize.Width / glControl1.ClientSize.Height;
             glControl1.Camera.Update();
         }
